% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lite_db.R
\name{append_new_records.data.frame}
\alias{append_new_records.data.frame}
\title{Append new df records to an existing table in a sql lite db}
\usage{
\method{append_new_records}{data.frame}(df, conn, db_name, tbl_name, key)
}
\arguments{
\item{key}{}
}
\description{
Append new df records to an existing table in a sql lite db
}
