% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lite_db.R
\name{append_new_records}
\alias{append_new_records}
\title{Append new records to an existing table in a sql lite db}
\usage{
append_new_records(x, ...)
}
\arguments{
\item{df}{data.frame or sf}

\item{key}{unique ID to use to identify duplicates if appending to existing table: defaults to ALL columns}
}
\value{
number of records added
}
\description{
Append new records to an existing table in a sql lite db
}
\examples{
\dontrun{
shp_nc <- sf::st_read(system.file("shape/nc.shp", package="sf"))
conn <- dbConnect(RSQLite::SQLite(), dbname  = here::here('inst','extdata', 'nc_test.db'))
shp_nc_rnd <- shp_nc \%>\% mutate(across(where(is.numeric),  ~ rnorm(length(.),mean = 100))) # random values so these are new records
append_new_records(shp_nc_rnd \%>\% head(10),
                  conn,
                  db_name =  here::here('inst','extdata', 'nc_test.db') ,
                  tbl_name = 'nc_tbl',
                  key=NULL) # use all columns + ignore geometry

}

}
