% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lite_db.R
\name{append_new_records.sf}
\alias{append_new_records.sf}
\title{Append new sf records to an existing table in a sql lite db}
\usage{
\method{append_new_records}{sf}(df, conn, db_name, tbl_name, key)
}
\arguments{
\item{key}{}
}
\description{
Append new sf records to an existing table in a sql lite db
}
