% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lite_db.R
\name{delete_spatial_obj}
\alias{delete_spatial_obj}
\title{Try to delete the auxiliary spatial data contained in a spatial lite db}
\usage{
delete_spatial_obj(conn, tbl_name)
}
\arguments{
\item{tbl_name}{}
}
\description{
Can be important because we might get errors if we delete only
a table with associated geometry without updating the geometry_columns table
}
