% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lite_db.R
\name{get_df_tbl.character}
\alias{get_df_tbl.character}
\title{Return a df using a db name}
\usage{
\method{get_df_tbl}{character}(db_name, tbl_name)
}
\arguments{
\item{tbl_name}{}
}
\description{
Return a df using a db name
}
