% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lite_db.R
\name{get_df_tbl}
\alias{get_df_tbl}
\title{Return a df from a sql lite connection that can be queried lazily with dplyr}
\usage{
get_df_tbl(x, ...)
}
\arguments{
\item{db_name}{or RSQLite::SQLiteConnection object}

\item{tbl_name}{}
}
\description{
Return a df from a sql lite connection that can be queried lazily with dplyr
}
