% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lite_db.R
\name{get_df_tbl.SQLiteConnection}
\alias{get_df_tbl.SQLiteConnection}
\title{Return a df using a SQllite connection}
\usage{
\method{get_df_tbl}{SQLiteConnection}(conn, tbl_name)
}
\arguments{
\item{tbl_name}{}
}
\description{
Return a df using a SQllite connection
}
