% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lite_db.R
\name{get_df_tbl_dispatch.character}
\alias{get_df_tbl_dispatch.character}
\title{Return a df using a db name}
\usage{
get_df_tbl_dispatch.character(
  db_name = here::here("inst", "extdata", "revue_technique.db"),
  tbl_name = "policies_2015-2019"
)
}
\arguments{
\item{tbl_name}{}
}
\description{
Return a df using a db name
}
