% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lite_db.R
\name{get_df_tbl_dispatch.SQLiteConnection}
\alias{get_df_tbl_dispatch.SQLiteConnection}
\title{Return a df using a SQllite connection}
\usage{
get_df_tbl_dispatch.SQLiteConnection(conn, tbl_name = "policies_2015-2019")
}
\arguments{
\item{tbl_name}{}
}
\description{
Return a df using a SQllite connection
}
