% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lite_db.R
\name{list_tables_db}
\alias{list_tables_db}
\title{Safely list existing tables in db}
\usage{
list_tables_db(conn)
}
\arguments{
\item{conn}{}
}
\description{
Will nt fail if DB does not exist, it will just return no tables
}
