% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lite_db.R
\name{tbl_exists.character}
\alias{tbl_exists.character}
\title{Determine if table exists in db (using db name)}
\usage{
\method{tbl_exists}{character}(db_name, tbl)
}
\arguments{
\item{tbl}{}
}
\value{
T if table exists
}
\description{
Determine if table exists in db (using db name)
}
