% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lite_db.R
\name{tbl_exists.SQLiteConnection}
\alias{tbl_exists.SQLiteConnection}
\title{Determine if table exists in db (using SQL lite connection)}
\usage{
\method{tbl_exists}{SQLiteConnection}(conn, tbl)
}
\arguments{
\item{tbl}{}
}
\value{
T if table exists
}
\description{
Determine if table exists in db (using SQL lite connection)
}
