% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lite_db.R
\name{write_table.data.frame}
\alias{write_table.data.frame}
\title{write_table for data.frame object}
\usage{
\method{write_table}{data.frame}(df, tbl_name, db_name, key = NULL, overwrite = F)
}
\arguments{
\item{overwrite}{}
}
\description{
write_table for data.frame object
}
