% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lite_db.R
\name{write_table}
\alias{write_table}
\title{Add a table to a SQLite db (or append if it exists)}
\usage{
write_table(df, ...)
}
\arguments{
\item{df}{data.frame or sf}

\item{key}{unique ID to use to identify duplicates if appending to existing table: defaults to ALL columns}

\item{overwrite}{}
}
\description{
Add a table to a SQLite db (or append if it exists)
}
\examples{
\dontrun{
 write_table  (df = iris,
tbl_name = 'iris_test_tbl',
key = NULL,
db_name = here::here('inst','extdata', 'iris_test.db'),
overwrite= T)
}
}
