% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lite_db.R
\name{write_table.sf}
\alias{write_table.sf}
\title{write_table for sf object}
\usage{
\method{write_table}{sf}(df, tbl_name, db_name, key = NULL, overwrite = F)
}
\arguments{
\item{overwrite}{}
}
\description{
write_table for sf object
}
