% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lite_db.R
\name{delete_tables}
\alias{delete_tables}
\title{Delete tables in a SQLite DB (works with all types of data - regular df or spatially indexed data)}
\usage{
delete_tables(db_name, list_tables)
}
\arguments{
\item{list_tables}{}
}
\value{
list_deleted_tables
}
\description{
Delete tables in a SQLite DB (works with all types of data - regular df or spatially indexed data)
}
